feat: escape filenames on Unix when using the shell option
Unescaped filenames could be used for command injection, because the `shell` option runs the tasks using `/bin/sh` or `cmd.exe`, depending on the platform.

This commit adds escaping of the files when running on an Unix shell, by escaping everything expect alphanumeric characters or the underscore with a backslash.

Windows escaping has not been solved yet, and should be done in a later commit.